fix the build from the header. @markcfong owes me donuts...
