+++ tensor_detector/CMakeLists.txt
@@ -41,7 +41,7 @@ target_include_directories(detector PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>)
 target_compile_features(detector PUBLIC c_std_99 cxx_std_17)  # Require C99 and C++17
-target_link_libraries(detector nvinfer nvinfer_plugin nvcaffe_parser nvonnxparser)
+target_link_libraries(detector nvinfer nvinfer_plugin)
 ament_target_dependencies(detector OpenCV)
 
 cuda_add_executable(converter src/converter.cpp)
